refactor(billing): address PR review comments
- Move hooks to web/src/hooks/ (useBillingInformation, useLicense)
- Rename actions.ts to svc.ts for consistency
- Use getFormattedDateTime for date formatting in SubscriptionSummary
- Update all related imports
