+++ web/src/app/ee/admin/billing/SubscriptionSummary.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { InfoItem } from "./InfoItem";
 import { statusToDisplay, BillingInformation } from "@/lib/billing/utils";
+import { formatDateShort } from "@/lib/dateUtils";
 
 interface SubscriptionSummaryProps {
   billingInformation: BillingInformation;
@@ -9,11 +10,6 @@ interface SubscriptionSummaryProps {
 export function SubscriptionSummary({
   billingInformation,
 }: SubscriptionSummaryProps) {
-  const formatDate = (dateStr: string | null): string => {
-    if (!dateStr) return "—";
-    return new Date(dateStr).toLocaleDateString();
-  };
-
   return (
     <div className="grid grid-cols-2 gap-4">
       <InfoItem
@@ -26,11 +22,11 @@ export function SubscriptionSummary({
       />
       <InfoItem
         title="Billing Start"
-        value={formatDate(billingInformation.current_period_start)}
+        value={formatDateShort(billingInformation.current_period_start)}
       />
       <InfoItem
         title="Billing End"
-        value={formatDate(billingInformation.current_period_end)}
+        value={formatDateShort(billingInformation.current_period_end)}
       />
     </div>
   );
