+++ web/src/hooks/useLicense.ts
